Giving gene ontology its own function
